W02 Prepare: EJS Templates Activity
Overview
You'll be working on using EJS templates to build a working server-side app in this activity.

Instructions
Complete the following steps:

Download and Implement Server CodeLinks to an external site.
Review the Starter FilesLinks to an external site.
Build the applicationLinks to an external site.
EJS View EngineLinks to an external site.
Implement index route with Express and EJSLinks to an external site.
Create a Render.com AccountLinks to an external site.
Add a render.com application serviceLinks to an external site.
DATABASE LOGIN INFO Download DATABASE LOGIN INFOfile. You will need this during this step.
Deploy updates and test the applicationLinks to an external site.