Assignment 4 submission
Objective 1
All views meet the standards found in the Frontend Checklist
The add inventory form is sticky when server-side checks return errors.
Obj. 1 value: Refer to grade book
Objective 2
Each of the three processes: 1) delivery of the management view, 2) processing of the new classification, and 3) processing of the new inventory item work correctly and meet the requirements listed in the tasks.
Messages are passed correctly through the session and displayed appropriately in views.
Obj. 2 value: Refer to grade book
Objective 3
Each of the task processes is implemented using the MVC architecture.
Each element of the MVC performs its proper function and is fully operational:
Routes function appropriately and call the controller to carry out the task.
Error handling middleware has been added to new routes.
Server-side validation middleware has been added to insertion routes.
The controller contains all logic for the application.
The model contains the database functionality and returns outcome information to the controller.
The classification select list is built using a custom function within the utilities > index file.
The views display the needed information for the human visitor.
Obj. 3 value: Refer to grade book
Objective 4
The functions in the model successfully insert data to the database using parameterized statements.
The functions in the model successfully report the outcome of the insertion back to the controller.
Obj. 4 value: Refer to grade book
Objective 5
Client-side validation is present and operational for all inputs in both the "add classification" and "add inventory" forms.
Server-side validation is present and operational for all inputs from the "add classification" and "add inventory" forms.
Correct data types for incoming data have been implemented.
Errors are successfully detected and returned to the appropriate view for correction.
Obj. 5 value: Refer to grade book
Objective 6
The render.com production URL and the GitHub repository URL are both submitted on time and correctly. Both URL's must be present and operational for the assignment to be graded.
Obj. 6 value: Refer to grade book